backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2

# ASR and audio utilities
faster-whisper==1.0.3
webrtcvad==2.0.10
soundfile==0.12.1
numpy==1.26.4

# Database and ORM
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
asyncpg==0.30.0
alembic==1.16.4

# Vector search (start with CPU)
faiss-cpu==1.8.0.post1

# Embeddings for retrieval (Phase 3)
sentence-transformers==3.0.1

# Optional utilities
httpx==0.27.0
